% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Counting Missing Values}
\usage{
na_count(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
Numerical vector
}
\description{
Counting Missing Values
}
\examples{
na_count(c(1, 2, NA, NA))
}
