% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Count missing values}
\usage{
na_count(x)
}
\arguments{
\item{x}{A vector or list}
}
\value{
A numerical vector
}
\description{
Count missing values
}
\examples{
na_count(c(1, 2, NA, NA))


}
