% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dotted.R
\name{theme_dotted}
\alias{theme_dotted}
\title{Theme dotted - custom ggplot theme}
\usage{
theme_dotted()
}
\value{
A theme function
}
\description{
Theme dotted - custom ggplot theme
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(mpg,hp)) + geom_point()
p + theme_dotted()

}
