% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dotted.R
\name{theme_dotted}
\alias{theme_dotted}
\title{Theme Dotted - a custom ggplot2 theme}
\usage{
theme_dotted()
}
\value{
A theme function
}
\description{
Theme Dotted - a custom ggplot2 theme
}
\examples{
library(ggplot2)

p <- ggplot2::ggplot(mtcars, aes(mpg, hp)) + ggplot2::geom_point()
p + theme_dotted()
}
